backend/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "frank-dannie",email = "frank.dannie@gmail.com"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "imagehash (>=4.3.2,<5.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "pydantic (>=2.12.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
